Added Cantilever.py
Added the Cantilever.py file which runs a FE model of the cantilever beam under gravity loading.

Also changed some of the parameter optimisation script to be more organised and started writing the objective function for minimising the error between data and points on the surface of the FE model.