Worked on Objective Function
Added the visualise.com script which will allow CMgui to work. Then also added a lot to the project_surface_data.py script, mostly working on the objective function.

This work included the code to select only the relevant elements on the surface of the FE model to use as candidates for projection, and also tried to select the faces, but the number which corresponds to each face isn't known, so this is likely to be wrong. Also another function is started which will extract the data node locations from the file and place them in a data variable ready to be passed to the objective function.