Created new file for simulation class and objective function
Added CantileverSimulation.py containing the CantileverSimulation class. This comes with some functions to set the parameters for the FE cantilever simulation and one (large function) which will prepare the simulation.

Outside the class, but in the same file, are several functions to solve the model, export data and an (empty) objective function which will hopefully contain an instance of the model, solve it repeatedly and inspect its output for optimisation of the parameters.