(Possibly) Finished CantileverSimulation Class
Added several of the helper functions to be methods of the CantileverSimulation class, including the solving, exporting, projection preparation and projection calculation functions.

The objective function is still independent of the simulation, and gets a setup simulation passed to it before altering the material properties and then finding the error.