Testing .gitignore
The problem may have been that there was no * before the file types, or that the file types were not what I was using.

This was successful in ignoring the output from the cantilever simulation. Now I need to find a way to make it stay only when I need to keep it for the unit testing.