Testing Boundary Condition Automation
Have set up frame for testing function, but need to find some way of inspecting what the BC is for the nodes so that it can be checked.

Have written code to automate selection of nodes on y=0 face of cantilever, but now having trouble with the convergence of the model if the number of elements is changed.