Fixed Convergence Bug
The problem wasn't converging because the BCs weren't being set properly by the automated code.

Changed the calculation for the maximum number of nodes in the simulation by adding one to each direction's number of elements (since there are n+1 nodes in each direction for n elements). This results in setting the BCs correctly.

Still can't test this against the unit test function because there is no function with which I can inspect the boundary condition which has been applied to a specific node.