Creating Test Dataset
So far have only set up the test data set, though I realise that no more than this may be required.

The test script hasn't been written yet because I don't know how the datapoint's locations will be stored, and so how they should be compared. I have decided that the test data will only look at the four corners of the model under zero load, so that these locations should be easily determinable analytically.

I am currently having trouble with the CMISS function which is supposed to evaluate the position of a point by interpolating the xi values. The problem I am having is that the function isn't reading the array of xi values correctly.