Worked on Optimisation script
Tried to get the optimisation routine working since I couldn't make any progress with either the BC automation or the creation of artificial data.

Now I'm stuck with several errors around getting the objective function to be stored within the parameterEstimation object so it can be supplied to the optimising routine.