Optimisation is Functional
The optimisation script is now functional. It produces two parameters as its output. It needs to be changed so that the second parameter is fixed (which doesn't work if the second parameter is set to zero every call, it needs to be removed from the inputs).

In addition, the Data creation function closer to working, and it is now accepting the array of xi values.