Optimisation Working
The optimisation routine is working and the second parameter has been fixed by simply calling it constantly with a value of zero. This produces output values for the second parameter in the order of e-5 to e-8 which seems acceptably small.

The Boundary Condition automation is still working, but without the unit test, and the data creation function is also working, though the unit testing is not running.