Mesh Convergence is working
The mesh convergence script is now working, but the simulations are taking a very long time to solve. The current one is 50x40x50 elements and is taking a very long time to solve.

Once it is done, if another one needs to be run, I will instead edit the calculation which increases the number of elements in each direction so it isn't so aggressive and instead eases the number of elements up more slowly, which will hopefully avoid having to run two very high detail models.