Improved Error checks
I improved the way the script assesses the error in a simulation. Now it find the maximum error between points in each axial direction and selects the largest of these to compare to the tolerance. For the overall error, it checks each point in each direction, and if any of these are above the tolerance it will perform another iteration.

There is an error which keeps occuring whenever I try to solve a model which has 8x5x5 elements or more in it: "onlinear line search solver did not converge. PETSc diverged line search". I'll get Prasad's help on this.