Working Script
The script runs through the loops, except it throws an error when a certain angle is reached due to trying to square root a negative number. I think that this will start occuring when the angles are such that the sin terms start producing negative numbers.

To get around it, I'll implement some conditional statements which perform the maths on the absolute values, but then multiply by negative one if required.