Calculations seem to work, now divergence error
The maths seems to be fine with calculating the gravity vector now, but the script throws a divergence error.

Interestingly, this occurs when the angles reach (0, 0), which is the 'resting' or unchanged orientation. I don't know how to fix this error, but I have a feeling its about the size of the steps which are being taken, so the simulations should be done in parts and the previous solution fed in as the starting ponit of the next one. Prasad should have this code.