Convergence tests working
The convergence script runs properly now and I have tested it with very coarse tolerances.

The problem now is that I don't know how fine the mesh needs to be to reach good tolerance levels (10^-3 to 10^-6). Such a fine mesh could take up so much RAM that it crashes the HPC computer, which would be bad. So I'm going to need to have another chat with Prasad as to how I should find out how much memory the calculations need to estimate such sizes. I should also have a think about how to find the mesh size in other ways in case this won't work.