Sensitivity Script Working
The sensitivity script now runs without throwing any errors. It doesn't produce the intended results though.

Firstly, there are a few optimisation calls which do not converge and so have their results replaced with NaN's.

In addition, the Hessian results for the other positions are all either 1 or -1, which isn't correct. I'll have a look into why this is occuring since the normal ParameterOptimsation script produces sensible results.