Fixed Hessian Calculations
The Hessian values are now being calculated correctly. I still don't understand what the divergence error is which is effecting some of the results (these are currently being replaced with NaN's).

In addition to that error, I also need to write a section which will plot the results, which doesn't work right now since I don't have the Tkinter file which matplotlib needs.