Projections Working
The data points are generated and now the data projections are also working.

There is a problem, however, with the faces onto which the projections are being made. Currently all the data points are being projected onto the end face, but this needs to be changed depending on which element it is. The issue is that the corner elements need to be able to have projections onto multiple faces, but I don't know if this is allowed.