Finished New Projection Script
Finished writing up a new projection script which takes care of both setting up each of the five projections and then doing the calculations all inside a single function call. This can be placed inside the objective function to evaluate the error after solving the FE model if the data set has already been obtained prior to calling the objective function.

This does need to be tested thoroughly to iron out all the bugs as I expect it to fail.