Working Sensitivity Script plus visualisation
The Optimal Design script works completely now. Given an assumedly true parameter value, the script will set up a simulation and loop through angles. At each one, it will perturb the parameter slightly allowing calculation of the Hessian which can then be used to determine which orientation provides the maximum identifyability.

In addition, this data is exported and then used to create a plot by running a separate function in a separate project.